Moved Encoder Method to Robot methods class.
