+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot.java
@@ -1,7 +1,9 @@
 package org.firstinspires.ftc.teamcode;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
+@Disabled
 public class Robot {
 
     DcMotor frontLeft;
@@ -31,6 +33,51 @@ public class Robot {
         frontRight.setPower(0);
         backRight.setPower(0);
         backLeft.setPower(0);
+    }
+    public void DriveForwardDistance(int inches, int power){
+        int diameter = 1;
+
+        //Rest Encoders.
+        frontLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        frontRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+
+        //Calculate tick value that the motors need to turn.
+        double circumference = 3.14*diameter;
+        double rotationsNeeded = inches/circumference;
+        int encoderDrivingTarget = (int)(rotationsNeeded*1120);
+
+        //Set tick value to the motors' target position.
+        frontLeft.setTargetPosition(encoderDrivingTarget);
+        frontRight.setTargetPosition(encoderDrivingTarget);
+        backLeft.setTargetPosition(encoderDrivingTarget);
+        backRight.setTargetPosition(encoderDrivingTarget);
+
+        //Set the desired power
+        frontLeft.setPower(power);
+        frontRight.setPower(power);
+        backLeft.setPower(power);
+        backRight.setPower(power);
+
+        //Set the mode of the motors to run to the target position.
+        frontLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        frontRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        backLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        backRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+        while(frontLeft.isBusy() || frontRight.isBusy() || backLeft.isBusy() || backRight.isBusy()){
+            telemetry.addData("Satus", "Driving");
+            telemetry.update();
+        }
+
+        frontLeft.setPower(0);
+        frontRight.setPower(0);
+        backLeft.setPower(0);
+        backRight.setPower(0);
+
+        telemetry.addData("Status", "Driving Complete");
+        telemetry.update();
     }
